# IRGP/openapi/instant-report-platform-openapi.yaml
openapi: 3.1.0
info:
  title: Instant Report Generation Platform API
  version: 1.0.0
  description: OpenAPI contract for the Instant Report Generation Platform services.
servers:
  - url: http://localhost:8000
    description: Default service gateway
paths:
...
  /api/v1/organizations/{organization_id}/subscription:
    get:
      summary: Get current subscription
      tags: [Subscriptions]
      parameters:
        - in: path
          name: organization_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current subscription status
  /api/v1/organizations/{organization_id}/subscription/upgrade:
    post:
      summary: Upgrade organization subscription
      tags: [Subscriptions]
      parameters:
        - in: path
          name: organization_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subscription upgraded
...
  /api/v1/notifications/{notification_id}/read:
    patch:
      summary: Mark notification as read
      tags: [Notifications]
      parameters:
        - in: path
          name: notification_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Notification read update
...
  /api/v1/reports/{report_id}:
    get:
      summary: Retrieve report by id
      tags: [Reports]
      parameters:
        - in: path
          name: report_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Report returned
  /api/v1/reports/jobs/{job_id}/status:
    get:
      summary: Get report job status
      tags: [Reports]
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job status
...